moving class definitions to own files
